vector tile work - moving to separate file so we can pull in to react-leaflet
